+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Feat
 
+- Allow specifying PIC version (#266)
+- Use arm64 pic binaries (#268)
 - **pic**: add costSchedule option and bump PocketIC to v13 (#264)
 
 ## 0.21.0 (2026-03-18)