+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Feat
 
-- **pic**: bump PocketIC to v14, add disableIngressValidation and senderInfo options
+- Update to v14, support disableIngressValidation and senderInfo (#271)
+- Allow specifying PIC version (#266)
+- Use arm64 pic binaries (#268)
 - **pic**: add costSchedule option and bump PocketIC to v13 (#264)
 
 ## 0.21.0 (2026-03-18)